feat: enforce browser fetch rules in CORS checker
